+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # CodeNarc Change Log
 
-Version 1.3
+TODO: Version 1.3
 --------------------------------------
 New Rules
  - #359: **ClassEndsWithBlankLine** rule: Check whether the class ends with a blank line.
  - #362: **ClassStartsWithBlankLine** rule: Check whether the class starts with a blank line.
+
+Bug Fixes
+ - #364: **UnnecessarySetter** rule: Fix StringIndexOutOfBoundsException for single-letter property names (e.g. account.setE(3)).
 
  
 Version 1.2.1 (Aug 2018)